Split the distribution generation test files out into separate files for distribution generation and random number generation.
