+++ include/distribution_generation_tests.hpp
@@ -6,11 +6,6 @@
 #ifndef DISTRIBUTION_GENERATION_TESTS_HPP_INCLUDED
 #define DISTRIBUTION_GENERATION_TESTS_HPP_INCLUDED 1
 
-double g_uniform_prng(void);
-
-double mean(double *array,
-            int num_elements);
-
 double mean(double3 *array,
             int num_elements);
 
@@ -19,9 +14,6 @@ double mean_norm(double3 *array,
 
 double mean_modified_radius(double3 *pos,
                             int num_elements);
-
-double std_dev(double *array,
-               int num_elements);
 
 double std_dev(double3 *array,
                int num_elements);
